switched codes between live and billboard pages
